requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
google-generativeai==0.3.2
python-dotenv==1.0.0
channels==4.0.0
daphne==4.0.0
